requirements.txt
azure-core==1.30.1
azure-storage-blob==12.19.1
Flask==3.0.2
numpy==1.26.4